Add files via upload
This project delves into a dataset covering health and economic indicators for Eastern European countries from 1997 to 2021. Our primary objectives are twofold: to determine the factors influencing child mortality rates in the region and to predict whether a country is predominantly urban or rural based on various health and economic indicators.

We start with a preliminary analysis of the dataset, examining the available features. Following this, we clean the data to ensure its integrity, addressing any missing or inconsistent values. An exploratory data analysis (EDA) is then conducted to gain deeper insights into the feature space, including distribution and density analyses.

For modeling, we employ multiple linear regression and LASSO regression to predict child mortality rates based on the dataset features. Additionally, we use logistic regression to address the question of whether a country is majority urban or rural based on the indicators.

This project offers a comprehensive analysis of health and economic indicators in Eastern Europe, providing valuable insights into the region's health outcomes and demographic characteristics.